requirements.txt
absl-py>=0.7.0
gym==0.17.3
matplotlib>=3.1.1
numpy
protobuf==3.17.3
pyserial==3.5
pybullet>=3.0.4
scikit-image>=0.17.2
scikit-learn>=0.23.2
shapely==1.7.1
xmltodict==0.12.0
